Added limit checks to the scaled X/Y value. Illegal scaled values were causing Xlib routines to consume almost 100% of the CPU.

+++ baselib/xygraph.h
@@ -989,8 +989,8 @@ void initPlotInfo (
 
 void addPoint (
   double oneX,
-  short scaledX,
-  short scaledY,
+  double scaledX,
+  double scaledY,
   int trace
 );
